+++ examples/plot_boston.py
@@ -53,9 +53,24 @@ boston_df = boston_df.set_index(list(boston_above_avg.columns))
 # Also give us access to the target (median house value)
 boston_df = boston_df.assign(median_value=boston.target)
 
+##########################################################################
+
 # UpSet plot it!
 upset = UpSet(boston_df, subset_size='count', intersection_plot_elements=3)
 upset.add_catplot(value='median_value', kind='strip', color='blue')
 upset.add_catplot(value='AGE', kind='strip', color='black')
 upset.plot()
+plt.title("UpSet with catplots, for orientation='horizontal'")
+plt.show()
+
+##########################################################################
+
+# And again in vertical orientation
+
+upset = UpSet(boston_df, subset_size='count', intersection_plot_elements=3,
+              orientation='vertical')
+upset.add_catplot(value='median_value', kind='strip', color='blue')
+upset.add_catplot(value='AGE', kind='strip', color='black')
+upset.plot()
+plt.title("UpSet with catplots, for orientation='vertical'")
 plt.show()