+++ upsetplot/tests/test_examples.py
@@ -1,6 +1,7 @@
 import glob
 import os
 import subprocess
+import sys
 
 import pytest
 
@@ -13,4 +14,6 @@ exa_glob = os.path.join(os.path.dirname(os.path.abspath(__file__)),
 def test_example(path):
     pytest.importorskip('sklearn')
     pytest.importorskip('seaborn')
-    subprocess.check_output(['python', path])
+    env = os.environ.copy()
+    env["PYTHONPATH"] = os.getcwd() + ":" + env.get("PYTHONPATH", "")
+    subprocess.check_output([sys.executable, path], env=env)